Zentri code for secure mesh demo - MQTT talks to Azure
This is the November code that runs the secure mesh demo - the problems with
the serial port have been resolved.  Zentri polls the serial port for any UART data
from the Mesh, and keeps its UART buffer from getting full and then asserting RTS
(stopping Mesh from sending more data - it would then hang!).
This code has the regular MQTT libraries that talk to Azure (instead of the Azure
SDK, which wasn't being updated).  
NOTE!!!  The Zentri dev boards try to pulse GPIO 23 for the WLAN LED - sadly,
this also messes with our RTS line.  To eliminate this problem on the dev boards,
run the commands:
set system.indicator.gpio wlan -1
save
reboot
from the Zentri command line (need a pirahna board).  This shouldn't be a problem
on a real product, as we wouldn't design a board to do this.  Do we need to ask
Zentri if/how to programmiticatlly turn this off - or do we just need make sure not to
use GPIO 23 for our CTS??