Programatically detect  & change the setting for the wlan gpio setting.
At init time, read the value "system.indicator.gpio wlan", and if it is not set to -1,
then we set it to -1, and save it in NVRAM.  It is normally driving GPIO23, which interferes with our UART CTS line.  The board has to be rebooted after this code
runs before it takes affect.